Remove comments with design decisions from the backbone.toledo index.js file
A decision has been made concerning how to load the backbone.toledo lib. The comments might be interesting for future reference, but not necessatry for an actual build. The comments are still accessible in the git history. Since the lib seems to work fine now and the current solution seems rather clean, I'm not sure whether the information will be that relevant in the future.